コンポーネントに切り出し
